Add a newline at the end of test files
When running in some settings, a C compiler may demand newlines at the
end of each file.  Instead of modifying everywhere that writes out test
files to incorporate newlines in each indivudual string, simply add a
newline when writing it out.

An examples of when this is a problem is running with `CC=clang` and
`CFLAGS="--std=c99 -pedantic-errors"`